app/requirements.txt
flask==2.3.2
waitress==2.1.2
psycopg2-binary==2.9.6